requirements.txt
boto3
langchain
langchain-aws
streamlit
chardet
python-dotenv
langchain-community
